infra/cdk/requirements.txt
aws-cdk.core==1.20.0
aws-cdk.aws_dynamodb==1.20.0
aws-cdk.aws_dynamodb_global==1.20.0
aws-cdk.aws_lambda==1.20.0